Extracted KeyValueStore interface
